+++ src/main/java/org/launchcode/adventureland/models/Category.java
@@ -1,0 +1,38 @@
+package org.launchcode.adventureland.models;
+
+import javax.persistence.Entity;
+
+@Entity
+public class Category extends AbstractEntity{
+
+    private String name;
+    private String description;
+
+    public Category(){}
+
+    public Category(String name, String description){
+        this.name = name;
+        this.description = description;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
