+++ src/main/java/org/launchcode/adventureland/models/data/CategoryRepository.java
@@ -1,9 +1,9 @@
 package org.launchcode.adventureland.models.data;
 
-import org.launchcode.adventureland.models.Group;
+import org.launchcode.adventureland.models.Category;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends CrudRepository<Group, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
 }
