+++ src/main/resources/templates/categories/view.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org/">
+<head th:replace="fragments :: head"></head>
+
+<body class="container  body-content">
+<h1 th:text="${'Category: ' + category.name}">Default Title</h1>
+
+<nav th:replace="fragments :: page-header"></nav>
+<div>
+    <p th:text="${'Name: ' + category.name}"></p>
+    <p th:text="${'Category: ' + category.description}"></p>
+</div>
+</body>
+</html>
