--- src/main/resources/templates/group/view.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org/">
-<head>
-    <meta cahrset="UTF-8"/>
-    <title></title>
-</head>
-<body>
-<p>Hello Jojo!!!!</p>
-</body>
-</html>
